+++ exporter/prometheusexporter/go.mod
@@ -1,0 +1,17 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusexporter
+
+go 1.16
+
+require (
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal v0.0.0-00010101000000-000000000000
+	github.com/prometheus/client_golang v1.11.0
+	github.com/prometheus/client_model v0.2.0
+	github.com/prometheus/prometheus v1.8.2-0.20210621150501-ff58416a0b02
+	github.com/stretchr/testify v1.7.0
+	go.opentelemetry.io/collector v0.33.1-0.20210820002854-d3000232f8f6
+	go.opentelemetry.io/collector/model v0.33.1-0.20210820002854-d3000232f8f6
+	go.uber.org/zap v1.19.0
+	gopkg.in/yaml.v2 v2.4.0
+)
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal => ../../internal/coreinternal
