+++ exporter/prometheusremotewriteexporter/go.mod
@@ -1,0 +1,17 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/exporter/prometheusremotewriteexporter
+
+go 1.16
+
+require (
+	github.com/gogo/protobuf v1.3.2
+	github.com/golang/snappy v0.0.4
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal v0.0.0-00010101000000-000000000000
+	github.com/prometheus/common v0.30.0
+	github.com/prometheus/prometheus v1.8.2-0.20210621150501-ff58416a0b02
+	github.com/stretchr/testify v1.7.0
+	github.com/tidwall/wal v0.1.5
+	go.opentelemetry.io/collector v0.33.1-0.20210820002854-d3000232f8f6
+	go.opentelemetry.io/collector/model v0.33.1-0.20210820002854-d3000232f8f6
+)
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal => ../../internal/coreinternal
