+++ exporter/prometheusremotewriteexporter/helper.go
@@ -25,7 +25,6 @@ import (
 
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/prompb"
-
 	"go.opentelemetry.io/collector/model/pdata"
 )
