+++ exporter/prometheusremotewriteexporter/img/cortex.png
