+++ exporter/prometheusremotewriteexporter/img/timeseries.png
