+++ exporter/prometheusremotewriteexporter/testdata/negative_queue_size.yaml
