+++ exporter/prometheusremotewriteexporter/testutil_test.go
@@ -19,7 +19,6 @@ import (
 	"time"
 
 	"github.com/prometheus/prometheus/prompb"
-
 	"go.opentelemetry.io/collector/model/pdata"
 )
